Optimize decoding blocks of postings using the vector API. (#13636) (#13652)
Our postings use a layout that helps take advantage of Java's
auto-vectorization to be reasonably fast to decode. But we can make it a bit
faster by vectorizing directly from the MemorySegment instead of first
copying data into a long[].

This approach only works when the `Directory` uses `MemorySegmentIndexInput`
under the hood, ie. `MMapDirectory` on JDK 21+.

+++ lucene/core/src/java21/org/apache/lucene/store/MemorySegmentAccessInput.java
@@ -27,7 +27,7 @@ import java.lang.foreign.MemorySegment;
 public interface MemorySegmentAccessInput extends RandomAccessInput, Cloneable {
 
   /** Returns the memory segment for a given position and length, or null. */
-  MemorySegment segmentSliceOrNull(long pos, int len) throws IOException;
+  MemorySegment segmentSliceOrNull(long pos, long len) throws IOException;
 
   default void readBytes(long pos, byte[] bytes, int offset, int length) throws IOException {
     for (int i = 0; i < length; i++) {
